# internal/identity-svc/internal/adapters/inbound/http/dto/dto.gen.go
// Package dto provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.5.0 DO NOT EDIT.
package dto

import (
	"context"
	"fmt"
	"net/http"

	"github.com/go-chi/chi/v5"
	openapi_types "github.com/oapi-codegen/runtime/types"
)

const (
	BearerAuthScopes = "bearerAuth.Scopes"
)

// Defines values for CartStatus.
const (
	CartStatusActive     CartStatus = "active"
	CartStatusCheckedOut CartStatus = "checked_out"
	CartStatusExpired    CartStatus = "expired"
)

// Defines values for CheckoutErrorCode.
const (
	CARTEMPTY                                CheckoutErrorCode = "CART_EMPTY"
	CARTNOTFOUND                             CheckoutErrorCode = "CART_NOT_FOUND"
	CONFLICT                                 CheckoutErrorCode = "CONFLICT"
	IDEMPOTENCYKEYREUSEDWITHDIFFERENTPAYLOAD CheckoutErrorCode = "IDEMPOTENCY_KEY_REUSED_WITH_DIFFERENT_PAYLOAD"
	INTERNAL                                 CheckoutErrorCode = "INTERNAL"
	INVALIDARGUMENT                          CheckoutErrorCode = "INVALID_ARGUMENT"
	PAYMENTDECLINED                          CheckoutErrorCode = "PAYMENT_DECLINED"
	SKUNOTFOUND                              CheckoutErrorCode = "SKU_NOT_FOUND"
	STOCKUNAVAILABLE                         CheckoutErrorCode = "STOCK_UNAVAILABLE"
)

// Defines values for OrderStatus.
const (
	AwaitingPayment OrderStatus = "awaiting_payment"
	AwaitingStock   OrderStatus = "awaiting_stock"
	Cancelled       OrderStatus = "cancelled"
	Confirmed       OrderStatus = "confirmed"
	Pending         OrderStatus = "pending"
)

// Defines values for ProductStatus.
const (
	Archived  ProductStatus = "archived"
	Draft     ProductStatus = "draft"
	Published ProductStatus = "published"
)

// Defines values for UserProfileRole.
const (
	Admin UserProfileRole = "admin"
	User  UserProfileRole = "user"
)

// Defines values for UserProfileStatus.
const (
	UserProfileStatusActive   UserProfileStatus = "active"
	UserProfileStatusDisabled UserProfileStatus = "disabled"
)

// AddCartItemRequest defines model for AddCartItemRequest.
type AddCartItemRequest struct {
	Quantity int    `json:"quantity" validate:"required,min=1"`
	Sku      string `json:"sku" validate:"required"`
}

// AuthTokensResponse defines model for AuthTokensResponse.
type AuthTokensResponse struct {
	AccessToken  string `json:"accessToken"`
	RefreshToken string `json:"refreshToken"`
}

// Cart defines model for Cart.
type Cart struct {
	CartId      string     `json:"cartId"`
	Currency    string     `json:"currency"`
	Items       []CartItem `json:"items"`
	Status      CartStatus `json:"status"`
	TotalAmount int        `json:"totalAmount"`
	UserId      string     `json:"userId"`
}

// CartStatus defines model for Cart.Status.
type CartStatus string

// CartItem defines model for CartItem.
type CartItem struct {
	Currency  *string `json:"currency,omitempty"`
	LineTotal int     `json:"lineTotal"`
	Name      *string `json:"name,omitempty"`
	Quantity  int     `json:"quantity"`
	Sku       string  `json:"sku"`
	UnitPrice int     `json:"unitPrice"`
}

// CheckoutErrorCode Canonical checkout error code set mapped from ecommerce.order.v1.CheckoutErrorCode.
type CheckoutErrorCode string

// CheckoutErrorResponse defines model for CheckoutErrorResponse.
type CheckoutErrorResponse struct {
	// Code Canonical checkout error code set mapped from ecommerce.order.v1.CheckoutErrorCode.
	Code          CheckoutErrorCode `json:"code"`
	CorrelationId *string           `json:"correlationId,omitempty"`
	Message       string            `json:"message"`
}

// CreateOrderRequest defines model for CreateOrderRequest.
type CreateOrderRequest struct {
	// PaymentMethod Placeholder for initial payment method selection model.
	// Contract mapping: forwarded as-is to internal payment proto field `ecommerce.payment.v1.InitiatePaymentRequest.provider_name`.
	PaymentMethod *string `json:"paymentMethod,omitempty" validate:"omitempty,min=1,max=50"`
}

// ErrorResponse defines model for ErrorResponse.
type ErrorResponse struct {
	Code          string  `json:"code"`
	CorrelationId *string `json:"correlationId,omitempty"`
	Message       string  `json:"message"`
}

// LoginRequest defines model for LoginRequest.
type LoginRequest struct {
	Email    openapi_types.Email `json:"email" validate:"required,email,max=254"`
	Password string              `json:"password" validate:"required,min=1,max=72"`
}

// Order defines model for Order.
type Order struct {
	Currency    string      `json:"currency"`
	Items       []OrderItem `json:"items"`
	OrderId     string      `json:"orderId"`
	Status      OrderStatus `json:"status"`
	TotalAmount int         `json:"totalAmount"`
	UserId      string      `json:"userId"`
}

// OrderStatus defines model for Order.Status.
type OrderStatus string

// OrderItem defines model for OrderItem.
type OrderItem struct {
	Currency  *string `json:"currency,omitempty"`
	LineTotal int     `json:"lineTotal"`
	Name      *string `json:"name,omitempty"`
	ProductId *string `json:"productId,omitempty"`
	Quantity  int     `json:"quantity"`
	Sku       string  `json:"sku"`
	UnitPrice int     `json:"unitPrice"`
}

// Product defines model for Product.
type Product struct {
	CategoryId  *string       `json:"categoryId,omitempty"`
	Currency    string        `json:"currency"`
	Description *string       `json:"description,omitempty"`
	Name        string        `json:"name"`
	Price       int           `json:"price"`
	ProductId   string        `json:"productId"`
	Sku         string        `json:"sku"`
	Status      ProductStatus `json:"status"`
}

// ProductStatus defines model for Product.Status.
type ProductStatus string

// ProductListResponse defines model for ProductListResponse.
type ProductListResponse struct {
	Items []Product `json:"items"`
}

// RefreshTokenRequest defines model for RefreshTokenRequest.
type RefreshTokenRequest struct {
	RefreshToken string `json:"refreshToken" validate:"required"`
}

// RegisterRequest defines model for RegisterRequest.
type RegisterRequest struct {
	DisplayName *string             `json:"displayName,omitempty" validate:"omitempty,min=1,max=100"`
	Email       openapi_types.Email `json:"email" validate:"required,email,max=254"`
	Password    string              `json:"password" validate:"required,min=8,max=72"`
}

// UpdateCartItemRequest defines model for UpdateCartItemRequest.
type UpdateCartItemRequest struct {
	Quantity int `json:"quantity" validate:"required,min=1"`
}

// UpdateProfileRequest defines model for UpdateProfileRequest.
type UpdateProfileRequest struct {
	DisplayName *string `json:"displayName,omitempty" validate:"omitempty,min=1,max=100"`
}

// UserProfile defines model for UserProfile.
type UserProfile struct {
	DisplayName *string             `json:"displayName,omitempty"`
	Email       openapi_types.Email `json:"email"`
	Role        UserProfileRole     `json:"role"`
	Status      UserProfileStatus   `json:"status"`
	UserId      string              `json:"userId"`
}

// UserProfileRole defines model for UserProfile.Role.
type UserProfileRole string

// UserProfileStatus defines model for UserProfile.Status.
type UserProfileStatus string

// CorrelationId defines model for CorrelationId.
type CorrelationId = string

// IdempotencyKey defines model for IdempotencyKey.
type IdempotencyKey = string

// OrderId defines model for OrderId.
type OrderId = string

// ProductId defines model for ProductId.
type ProductId = string

// Sku defines model for Sku.
type Sku = string

// BadRequest defines model for BadRequest.
type BadRequest = ErrorResponse

// CheckoutBadRequest defines model for CheckoutBadRequest.
type CheckoutBadRequest = CheckoutErrorResponse

// CheckoutConflict defines model for CheckoutConflict.
type CheckoutConflict = CheckoutErrorResponse

// CheckoutInternalServerError defines model for CheckoutInternalServerError.
type CheckoutInternalServerError = CheckoutErrorResponse

// CheckoutNotFound defines model for CheckoutNotFound.
type CheckoutNotFound = CheckoutErrorResponse

// Conflict defines model for Conflict.
type Conflict = ErrorResponse

// Forbidden defines model for Forbidden.
type Forbidden = ErrorResponse

// NotFound defines model for NotFound.
type NotFound = ErrorResponse

// Unauthorized defines model for Unauthorized.
type Unauthorized = ErrorResponse

// LoginUserJSONRequestBody defines body for LoginUser for application/json ContentType.
type LoginUserJSONRequestBody = LoginRequest

// RefreshTokenJSONRequestBody defines body for RefreshToken for application/json ContentType.
type RefreshTokenJSONRequestBody = RefreshTokenRequest

// RegisterUserJSONRequestBody defines body for RegisterUser for application/json ContentType.
type RegisterUserJSONRequestBody = RegisterRequest

// RegisterAdminJSONRequestBody defines body for RegisterAdmin for application/json ContentType.
type RegisterAdminJSONRequestBody = RegisterRequest

// ServerInterface represents all server handlers.
type ServerInterface interface {
	// Authenticate user
	// (POST /v1/auth/login)
	LoginUser(w http.ResponseWriter, r *http.Request)
	// Refresh access token
	// (POST /v1/auth/refresh)
	RefreshToken(w http.ResponseWriter, r *http.Request)
	// Register a new user
	// (POST /v1/auth/register)
	RegisterUser(w http.ResponseWriter, r *http.Request)
	// Register a new admin user
	// (POST /v1/auth/register-admin)
	RegisterAdmin(w http.ResponseWriter, r *http.Request)
}

// Unimplemented server implementation that returns http.StatusNotImplemented for each endpoint.

type Unimplemented struct{}

// Authenticate user
// (POST /v1/auth/login)
func (_ Unimplemented) LoginUser(w http.ResponseWriter, r *http.Request) {
	w.WriteHeader(http.StatusNotImplemented)
}

// Refresh access token
// (POST /v1/auth/refresh)
func (_ Unimplemented) RefreshToken(w http.ResponseWriter, r *http.Request) {
	w.WriteHeader(http.StatusNotImplemented)
}

// Register a new user
// (POST /v1/auth/register)
func (_ Unimplemented) RegisterUser(w http.ResponseWriter, r *http.Request) {
	w.WriteHeader(http.StatusNotImplemented)
}

// Register a new admin user
// (POST /v1/auth/register-admin)
func (_ Unimplemented) RegisterAdmin(w http.ResponseWriter, r *http.Request) {
	w.WriteHeader(http.StatusNotImplemented)
}

// ServerInterfaceWrapper converts contexts to parameters.
type ServerInterfaceWrapper struct {
	Handler            ServerInterface
	HandlerMiddlewares []MiddlewareFunc
	ErrorHandlerFunc   func(w http.ResponseWriter, r *http.Request, err error)
}

type MiddlewareFunc func(http.Handler) http.Handler

// LoginUser operation middleware
func (siw *ServerInterfaceWrapper) LoginUser(w http.ResponseWriter, r *http.Request) {

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.LoginUser(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// RefreshToken operation middleware
func (siw *ServerInterfaceWrapper) RefreshToken(w http.ResponseWriter, r *http.Request) {

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.RefreshToken(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// RegisterUser operation middleware
func (siw *ServerInterfaceWrapper) RegisterUser(w http.ResponseWriter, r *http.Request) {

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.RegisterUser(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// RegisterAdmin operation middleware
func (siw *ServerInterfaceWrapper) RegisterAdmin(w http.ResponseWriter, r *http.Request) {

	ctx := r.Context()

	ctx = context.WithValue(ctx, BearerAuthScopes, []string{})

	r = r.WithContext(ctx)

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.RegisterAdmin(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

type UnescapedCookieParamError struct {
	ParamName string
	Err       error
}

func (e *UnescapedCookieParamError) Error() string {
	return fmt.Sprintf("error unescaping cookie parameter '%s'", e.ParamName)
}

func (e *UnescapedCookieParamError) Unwrap() error {
	return e.Err
}

type UnmarshalingParamError struct {
	ParamName string
	Err       error
}

func (e *UnmarshalingParamError) Error() string {
	return fmt.Sprintf("Error unmarshaling parameter %s as JSON: %s", e.ParamName, e.Err.Error())
}

func (e *UnmarshalingParamError) Unwrap() error {
	return e.Err
}

type RequiredParamError struct {
	ParamName string
}

func (e *RequiredParamError) Error() string {
	return fmt.Sprintf("Query argument %s is required, but not found", e.ParamName)
}

type RequiredHeaderError struct {
	ParamName string
	Err       error
}

func (e *RequiredHeaderError) Error() string {
	return fmt.Sprintf("Header parameter %s is required, but not found", e.ParamName)
}

func (e *RequiredHeaderError) Unwrap() error {
	return e.Err
}

type InvalidParamFormatError struct {
	ParamName string
	Err       error
}

func (e *InvalidParamFormatError) Error() string {
	return fmt.Sprintf("Invalid format for parameter %s: %s", e.ParamName, e.Err.Error())
}

func (e *InvalidParamFormatError) Unwrap() error {
	return e.Err
}

type TooManyValuesForParamError struct {
	ParamName string
	Count     int
}

func (e *TooManyValuesForParamError) Error() string {
	return fmt.Sprintf("Expected one value for %s, got %d", e.ParamName, e.Count)
}

// Handler creates http.Handler with routing matching OpenAPI spec.
func Handler(si ServerInterface) http.Handler {
	return HandlerWithOptions(si, ChiServerOptions{})
}

type ChiServerOptions struct {
	BaseURL          string
	BaseRouter       chi.Router
	Middlewares      []MiddlewareFunc
	ErrorHandlerFunc func(w http.ResponseWriter, r *http.Request, err error)
}

// HandlerFromMux creates http.Handler with routing matching OpenAPI spec based on the provided mux.
func HandlerFromMux(si ServerInterface, r chi.Router) http.Handler {
	return HandlerWithOptions(si, ChiServerOptions{
		BaseRouter: r,
	})
}

func HandlerFromMuxWithBaseURL(si ServerInterface, r chi.Router, baseURL string) http.Handler {
	return HandlerWithOptions(si, ChiServerOptions{
		BaseURL:    baseURL,
		BaseRouter: r,
	})
}

// HandlerWithOptions creates http.Handler with additional options
func HandlerWithOptions(si ServerInterface, options ChiServerOptions) http.Handler {
	r := options.BaseRouter

	if r == nil {
		r = chi.NewRouter()
	}
	if options.ErrorHandlerFunc == nil {
		options.ErrorHandlerFunc = func(w http.ResponseWriter, r *http.Request, err error) {
			http.Error(w, err.Error(), http.StatusBadRequest)
		}
	}
	wrapper := ServerInterfaceWrapper{
		Handler:            si,
		HandlerMiddlewares: options.Middlewares,
		ErrorHandlerFunc:   options.ErrorHandlerFunc,
	}

	r.Group(func(r chi.Router) {
		r.Post(options.BaseURL+"/v1/auth/login", wrapper.LoginUser)
	})
	r.Group(func(r chi.Router) {
		r.Post(options.BaseURL+"/v1/auth/refresh", wrapper.RefreshToken)
	})
	r.Group(func(r chi.Router) {
		r.Post(options.BaseURL+"/v1/auth/register", wrapper.RegisterUser)
	})
	r.Group(func(r chi.Router) {
		r.Post(options.BaseURL+"/v1/auth/register-admin", wrapper.RegisterAdmin)
	})

	return r
}